chore: rename cli-output into helpers and output
