+++ packages/respect-core/src/index.ts
@@ -3,6 +3,6 @@ export { run } from './run.js';
 export type { GenerateArazzoOptions } from './generate.js';
 export type { RespectOptions } from './run.js';
 export * from './types.js';
-export { maskSecrets } from './modules/cli-output/mask-secrets.js';
-export { calculateTotals } from './modules/cli-output/calculate-tests-passed.js';
-export { RESET_ESCAPE_CODE } from './utils/cli-outputs.js';
+export { maskSecrets } from './modules/output/mask-secrets.js';
+export { calculateTotals } from './modules/output/calculate-tests-passed.js';
+export { RESET_ESCAPE_CODE } from './modules/output/helpers.js';
