--- packages/respect-core/src/modules/cli-output/index.ts
@@ -1,6 +1,0 @@
-export * from './display-summary.js';
-export * from './display-errors.js';
-export * from './display-checks.js';
-export * from './calculate-tests-passed.js';
-export * from './verbose-logs.js';
-export * from './mask-secrets.js';
