+++ packages/respect-core/src/utils/ajv-errors.ts
@@ -1,5 +1,5 @@
 import betterAjvErrors from 'better-ajv-errors';
-import { RESET_ESCAPE_CODE } from './cli-outputs.js';
+import { RESET_ESCAPE_CODE } from '../modules/output/helpers.js';
 
 import type { JSONSchemaType } from '@redocly/ajv/dist/2020.js';
